Collect UIDs generically
