+++ yuml-rs/src/parser/utils.rs
@@ -1,4 +1,4 @@
-use crate::model::shared::LabeledElement;
+use crate::model::shared::{ElementDetails, LabeledElement};
 
 use super::*;
 
@@ -34,4 +34,34 @@ impl<'a, T: LabeledElement> Uids<'a, T> {
     pub fn get(&'a self, key: &str) -> Option<&'a (usize, &'a T)> {
         self.uids.get(key)
     }
+}
+
+pub fn populate_uids<T: LabeledElement>(elements: &[T]) -> (Uids<T>, Vec<ElementDetails<T>>) {
+    let mut uids = Uids::default();
+
+    // we must collect to borrow uids in subsequent iterator
+    let element_details: Vec<ElementDetails<T>> = elements
+        .iter()
+        .filter_map(|e| {
+            if e.is_connection() {
+                // ignore arrows for now
+                None
+            } else {
+                let lbl = e.label();
+                if uids.contains_key(&lbl) {
+                    None
+                } else {
+                    let id = uids.insert_uid(lbl, e);
+                    Some((id, e))
+                }
+            }
+        })
+        .map(|(id, element)| ElementDetails {
+            id: Some(id),
+            element,
+            relation: None,
+        })
+        .collect();
+
+    (uids, element_details)
 }
